move evaluateConstantSum.js
