+++ lib/collectAndCombine/addLikeTerms.js
@@ -2,7 +2,7 @@
 
 const clone = require('../util/clone');
 const CombineChecks = require('./CombineChecks');
-const evaluateConstantSum = require('../evaluateConstantSum');
+const evaluateConstantSum = require('./evaluateConstantSum');
 const MathChangeTypes = require('../MathChangeTypes');
 const NodeCreator = require('../util/NodeCreator');
 const NodeStatus = require('../util/NodeStatus');
