+++ lib/collectAndCombine/evaluateConstantSum.js
@@ -1,13 +1,13 @@
 'use strict';
 
-const addConstantFractions = require('./simplifyFractions/addConstantFractions');
-const addConstantAndFraction = require('./simplifyFractions/addConstantAndFraction');
-const clone = require('./util/clone');
-const evaluateArithmetic = require('./evaluateArithmetic');
-const MathChangeTypes = require('./MathChangeTypes');
-const NodeCreator = require('./util/NodeCreator');
-const NodeStatus = require('./util/NodeStatus');
-const NodeType = require('./util/NodeType');
+const evaluateArithmetic = require('../evaluateArithmetic');
+const MathChangeTypes = require('../MathChangeTypes');
+const addConstantFractions = require('../simplifyFractions/addConstantFractions');
+const addConstantAndFraction = require('../simplifyFractions/addConstantAndFraction');
+const clone = require('../util/clone');
+const NodeCreator = require('../util/NodeCreator');
+const NodeStatus = require('../util/NodeStatus');
+const NodeType = require('../util/NodeType');
 
 // Evaluates a sum of constant numbers and integer fractions to a single
 // constant number or integer fraction. e.g. e.g. 2/3 + 5 + 5/2 => 49/6
